Fixed wrong initialization bug, now it seems like the buffers are still overflowing because i'm summing coefficients which is illegal. I should accumulate ALL radiance values for a face of a cell, and after accumulation calculate the Monte Carlo estimated integral.

+++ mcrt-experiments/mcrt-experiments/LaunchParams.hpp
@@ -21,7 +21,7 @@ namespace mcrt {
 
 
 	struct SHWeights {
-		float* weights;
+		double* weights;
 		int size;
 		int amountBasisFunctions;
 	};
